BurgerBackend/requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
gunicorn==20.0.4
Flask-PyMongo==2.3.0
PyJWT==1.7.1
dnspython==1.16.0